layout[Prompt Input]:📦 Move prompt input from Dialog to '/new' route;
